pyproject.toml
[project]
name = "mtg-rules-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"


dependencies = [
    "faiss-cpu>=1.10.0",
    "langgraph>=0.2.73",
    "langchain-community>=0.3.17",
    "langchain>=0.3.18",
    "langchain-openai>=0.3.6",
    "python-dotenv>=1.0.1",
    "chainlit>=2.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "types-requests>=2.32.0.20241016",
    "tiktoken>=0.9.0",
    "requests>=2.32.3",
]


[tool.ruff.lint]
select = [
    "F",    # Pyflakes
    "I",    # isort
    "E",    # pycodestyle
    "S",    # flake8-bandit
    "N",    # pep8-naming
    "PERF",
]

ignore = [
    "N999", # Invalid module name
]
